% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_numbers.R
\name{deggape}
\alias{deggape}
\title{Title EGGAPE: Exponentiated Generalized Gull Alpha Power Exponential Distribution}
\usage{
deggape(x, a, b, alpha, lambda)
}
\arguments{
\item{x}{vector of quantities}

\item{a}{shape parameter of the distribution}

\item{b}{shape parameter of the distribution}

\item{alpha}{shape parameter of the distribution}

\item{lambda}{scale parameter of the distribution}
}
\value{
deggape gives the density function
}
\description{
Title EGGAPE: Exponentiated Generalized Gull Alpha Power Exponential Distribution
}
\examples{
deggape(c(seq(0,1, by = 0.1)), 1.5,1.5,2.5,0.2)
}
