% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_numbers.R
\name{qeggape}
\alias{qeggape}
\title{Title EGGAPE: Exponentiated Generalized Gull Alpha Power Exponential Distribution}
\usage{
qeggape(p, a, b, alpha, lambda)
}
\arguments{
\item{p}{Vector of probabilities}

\item{a}{Shape parameter}

\item{b}{Shape parameter}

\item{alpha}{Shape parameter}

\item{lambda}{Scale parameter}
}
\value{
numeric
}
\description{
Title EGGAPE: Exponentiated Generalized Gull Alpha Power Exponential Distribution
}
\examples{
qeggape(0.1,0.3,0.4,0.4,0.5)
}
