% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_numbers.R
\name{reggape}
\alias{reggape}
\title{Title EGGAPE: Exponentiated Generalized Gull Alpha Power Exponential Distribution}
\usage{
reggape(n, a, b, alpha, lambda)
}
\arguments{
\item{n}{Number of observations}

\item{a}{shape parameter of the distribution}

\item{b}{shape parameter of the distribution}

\item{alpha}{shape parameter of the distribution}

\item{lambda}{scale parameter of the distribution}
}
\value{
reggape generates random variates
}
\description{
Title EGGAPE: Exponentiated Generalized Gull Alpha Power Exponential Distribution
}
\examples{
reggape(100,1.5,1.8,0.4,0.5)
}
